Added signed casts

+++ mytar.c
@@ -157,7 +157,7 @@ void extract_file(const char *filename, size_t file_size, FILE *archive) {
     char buffer[512];
     int64_t bytes_remaining = file_size;
     while (bytes_remaining > 0) {
-        int64_t bytes_to_read = bytes_remaining < sizeof(buffer) ? bytes_remaining : sizeof(buffer);
+        int64_t bytes_to_read = bytes_remaining < (int64_t)sizeof(buffer) ? bytes_remaining : (int64_t)sizeof(buffer);
         int64_t bytes_read = fread(buffer, 1, sizeof(buffer), archive);
         if (bytes_read < bytes_to_read) {
             fprintf(stderr, "mytar: Unexpected EOF in archive\n");
